Move WorkerInfo to commons (#2899)
* Move WorkerInfo to commons

* add back worker dep
