+++ pulsar-client-tools/src/main/java/org/apache/pulsar/admin/cli/CmdFunctionWorker.java
@@ -30,7 +30,7 @@ import com.google.gson.JsonParser;
 
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.pulsar.functions.worker.WorkerInfo;
+import org.apache.pulsar.common.functions.WorkerInfo;
 
 import java.util.Collection;
 import java.util.List;
